+++ android/src/main/java/br/com/certiface/rn/sdk/model/Featues.kt
@@ -1,4 +1,4 @@
-package br.com.oititec.rn.sdk.model
+package br.com.certiface.rn.sdk.model
 
 enum class Features(label: String) {
   IProov("IProov"),