+++ react-native.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  dependency: {
+    platforms: {
+      android: {
+        packageName: 'br.com.certiface.rn.sdk',
+      },
+    },
+  },
+};